Add fonts config, improve builder

+++ bin/build_tool.py
@@ -4,36 +4,7 @@ import os
 import sys
 import argparse
 import shutil
-import subprocess
-import collections
-
-def mkdirs(directory):
-  if not os.path.exists(directory):
-    os.makedirs(directory)
-
-def show_file(fname):
-  with open(fname, 'r') as fin:
-    print fin.read()
-
-def call_verbose(action):
-  if isinstance(action, list): string = ' '.join(action)
-  else: string = action
-  print('Performing ' + string + ' ...\n')
-  try:
-    # Redirect stderr to stdout
-    if args.errors: log = subprocess.check_output(action)
-    else:           log = subprocess.check_output(action, stderr=subprocess.STDOUT)
-    show = []
-    interesting = ['Package name:', 'Version:', 'Date:', 'Status:', 'version',
-                   ' path.', ' path ']
-    for line in log.splitlines():
-      for interest in interesting:
-        if interest in line:
-          show.append(line)
-    print('\n'.join(list(collections.OrderedDict.fromkeys(show))))
-  except (subprocess.CalledProcessError, OSError) as e:
-    print("Execution of " + string + " failed:\n" + str(e) + '\n' + e.output)
-  print('\n... done!')
+from bcn_tools import *
 
 # Parse command line options
 parser = argparse.ArgumentParser(description='Build the Whizard')
@@ -74,15 +45,7 @@ parser.add_argument("-p", '--openmp', action='store_true',
 args = parser.parse_args()
 
 # Select a base path
-base_paths = ['~/trunk', '/data/bcho/trunk']
-for bpath in base_paths:
-  if os.path.exists(bpath):
-    base_path = bpath
-try:
-  print('base_path'.ljust(17) + '=\t' + base_path)
-except NameError:
-  print('No known base directory found!')
-  sys.exit(1)
+base_path = get_base_path()
 
 # Default compiler and optimization level
 #def set_default(option, value):
@@ -94,11 +57,11 @@ if not args.compiler:
 if not args.optimization:
   args.optimization = ['2']
 if not args.fcflags:
-  args.fcflags = ['-fmax-errors=1 -fbounds-check']
+  args.fcflags = ['-fmax-errors=1 -fbounds-check -Wall -Wuninitialized']
 if not args.configureflags:
   args.configureflags = ['']
 
-# Convenience builds
+# Convenience magic
 if 'ifort' in args.build:
   args.compiler = ['ifort']
   args.optimization = ['3']
@@ -107,18 +70,27 @@ if 'ifort' in args.build:
 if 'omp' in args.build:
   args.configureflags = ['--enable-fc-openmp']
 
+if 'dist' in args.build:
+  args.configureflags = ['--enable-distribution']
+
 # show set options for builder
 tasks = ['configure', 'autoreconf', 'make', 'makecheck', 'remove']
 options = ['jobs', 'errors']
 variants = ['configureflags', 'compiler', 'optimization', 'fcflags', 'build', 'tag']
 arg_dict = vars(args)
 for item in tasks + options + variants:
-  print(item.ljust(17) + '=\t' + str(arg_dict[item]))
+  show_variable(item, arg_dict[item])
+
+def _call_verbose(cmd):
+  lines_to_show = ['Package name:', 'Version:', 'Date:', 'Status:', 'version',
+                   ' path.', ' path ']
+  call_verbose(cmd, filter_strgs=lines_to_show, show_errors=args.errors)
+
 
 # autoreconf if desired
 os.chdir(base_path)
 if args.autoreconf:
-  call_verbose('autoreconf')
+  _call_verbose('autoreconf')
 
 # create build dirs
 build_name = args.build
@@ -126,7 +98,8 @@ if args.tag:
   build_name += '-' + args.tag
 build_path = os.path.join('build', build_name)
 if args.remove:
-  shutil.rmtree(build_path)
+  if os.path.isdir(build_path):
+    shutil.rmtree(build_path)
 mkdirs(build_path)
 os.chdir(build_path)
 
@@ -143,14 +116,14 @@ configure_options = [prefix, fortran_compiler, fortran_flags, configureflags]
 if args.configure:
   if compiler == 'ifort': package = os.path.join(base_path, 'omega')
   else:                   package = base_path
-  call_verbose([os.path.join(package, 'configure')] + configure_options)
+  _call_verbose([os.path.join(package, 'configure')] + configure_options)
 
 # build if desired
 if args.make:
-  call_verbose(['make', '-j' + args.jobs])
-  call_verbose(['make', '-j' + args.jobs, 'install'])
+  _call_verbose(['make', '-j' + str(args.jobs)])
+  _call_verbose(['make', '-j' + str(args.jobs), 'install'])
 
 # check if desired
 if args.makecheck:
-  call_verbose(['make', '-j' + args.jobs, 'check'])
+  _call_verbose(['make', '-j' + str(args.jobs), 'check'])
   show_file(os.path.join('tests', 'test-suite.log'))
